Sprint 2 increment 4: JSON API
* Refactor previous solution to pass the repo to the handlers.
* Use separate files for each handler.
